Add base mappings for each controller to fix ambiguity caused by different default mapping strategy on tomcat
