+++ src/main/java/nl/thehyve/ocdu/controllers/MappingController.java
@@ -41,31 +41,7 @@ public class MappingController {
         return new ResponseEntity<>(userItems,HttpStatus.OK);
     }
 
-    @RequestMapping(value = "/upload-mapping", method = RequestMethod.POST)
-    public ResponseEntity<List<OcItemMapping>> acceptMapping(HttpSession session, @RequestBody List<OcItemMapping> mappings) {
-        if (!isValid(mappings)) {
-            log.error("Incorrect mapping JSON provided.");
-            return new ResponseEntity<>(mappings, HttpStatus.BAD_REQUEST);
-        }
-        UploadSession submission = uploadSessionService.getCurrentUploadSession(session);
-        mappingService.applyMapping(mappings, submission);
-        return new ResponseEntity<>(mappings, HttpStatus.OK);
-    }
 
-    private boolean isValid(List<OcItemMapping> mappings) {
-        List<OcItemMapping> faulty = mappings.stream().filter(ocItemMapping -> {
-            if (ocItemMapping.getCrfName() == null ||
-                    ocItemMapping.getStudy() == null ||
-                    ocItemMapping.getCrfVersion() == null ||
-                    ocItemMapping.getEventName() == null ||
-                    ocItemMapping.getOcItemName() == null ||
-                    ocItemMapping.getUsrItemName() == null) return true;
-            else return false;
-        }).collect(Collectors.toList());
-        if (faulty.size() > 0) {
-            return false;
-        } else return true;
-    }
 }
