+++ src/main/java/nl/thehyve/ocdu/controllers/MetadataController.java
@@ -4,6 +4,8 @@ import nl.thehyve.ocdu.models.UploadSession;
 import nl.thehyve.ocdu.services.DataService;
 import nl.thehyve.ocdu.services.OcUserService;
 import nl.thehyve.ocdu.services.UploadSessionService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,7 +21,10 @@ import static org.springframework.http.HttpStatus.OK;
  * Created by piotrzakrzewski on 07/05/16.
  */
 @RestController
+@RequestMapping("/metadata")
 public class MetadataController {
+
+    private static final Logger log = LoggerFactory.getLogger(MetadataController.class);
 
     @Autowired
     DataService dataService;
@@ -30,18 +35,9 @@ public class MetadataController {
     @Autowired
     OcUserService ocUserService;
 
-    @RequestMapping(value = "/data-info", method = RequestMethod.GET)
-    public ResponseEntity<DataService.FieldsDetermined> getFieldsInfo(HttpSession session) {
-        UploadSession submission = uploadSessionService.getCurrentUploadSession(session);
-        DataService.FieldsDetermined info = dataService.getInfo(submission);
-        if (info == null) {
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        } else {
-            return new ResponseEntity<>(info, OK);
-        }
-    }
 
-    @RequestMapping(value = "/metadata-tree", method = RequestMethod.GET)
+
+    @RequestMapping(value = "/tree", method = RequestMethod.GET)
     public ResponseEntity<DataService.MetaDataTree> getMetadata(HttpSession session) {
         UploadSession currentUploadSession = uploadSessionService.getCurrentUploadSession(session);
         String pwd = ocUserService.getOcwsHash(session);
